+++ packages/common/src/Utils/Context.js
@@ -11,6 +11,7 @@ const RichContentContext = React.createContext({
   setEditorState: noop,
   locale: 'en',
   helpers: {},
+  languageDir: 'ltr',
 });
 const { Provider, Consumer } = RichContentContext;
 
@@ -26,6 +27,7 @@ Provider.propTypes = {
     helpers: PropTypes.object,
     config: PropTypes.object,
     isMobile: PropTypes.bool,
+    languageDir: PropTypes.string,
   }),
 };