+++ packages/editor/src/RichContentEditor/Toolbars/StaticToolbar/StaticToolbar.jsx
@@ -107,12 +107,12 @@ export default class StaticToolbar extends React.PureComponent {
       toolbarStyles.responsiveArrow
     );
     const leftArrowIconClassNames = classNames(
-      Styles.staticToolbar_responsiveArrowLeft_icon,
-      toolbarStyles.responsiveArrowLeft_icon
+      Styles.staticToolbar_responsiveArrowStart_icon,
+      toolbarStyles.responsiveArrowStart_icon
     );
     const rightArrowIconClassNames = classNames(
-      Styles.staticToolbar_responsiveArrowRight_icon,
-      toolbarStyles.responsiveArrowRight_icon
+      Styles.staticToolbar_responsiveArrowEnd_icon,
+      toolbarStyles.responsiveArrowEnd_icon
     );
 
     const buttonClassNames = classNames(Styles.staticToolbar_buttons, toolbarStyles.buttons);